Canary Deployment: Change - Make it extensible.
